Fixing compiler warnings

+++ NFIQ2/NFIQ2Algorithm/src/features/FingerJetFXFeature.cpp
@@ -28,8 +28,9 @@ NFIQ2::QualityFeatures::FingerJetFXFeature::centerOfMinutiaeMass(
 		lx += m.x;
 		ly += m.y;
 	}
-	return std::make_pair<unsigned int, unsigned int>(
-	    lx / minutiaData.size(), ly / minutiaData.size());
+	return std::make_pair<unsigned int, unsigned int>(lx /
+		minutiaData.size(),
+	    ly / minutiaData.size());
 }
 
 std::string
@@ -166,8 +167,8 @@ NFIQ2::QualityFeatures::FingerJetFXFeature::computeFeatureData(
 	}
 
 	unsigned int minCnt { 0 };
-	const FRFXLL_RESULT fxResMin = FRFXLLGetMinutiaInfo(
-	    hFeatureSet, &minCnt, nullptr);
+	const FRFXLL_RESULT fxResMin = FRFXLLGetMinutiaInfo(hFeatureSet,
+	    &minCnt, nullptr);
 	if (!FRFXLL_SUCCESS(fxResMin)) {
 		FRFXLLCloseHandle(&hFeatureSet);
 		throw NFIQ2::Exception(
@@ -185,8 +186,8 @@ NFIQ2::QualityFeatures::FingerJetFXFeature::computeFeatureData(
 		    "Could not allocate space for extracted minutiae records.");
 	}
 
-	const FRFXLL_RESULT fxResData = FRFXLLGetMinutiae(
-	    hFeatureSet, BASIC_19794_2_MINUTIA_STRUCT, &minCnt, mdata.get());
+	const FRFXLL_RESULT fxResData = FRFXLLGetMinutiae(hFeatureSet,
+	    BASIC_19794_2_MINUTIA_STRUCT, &minCnt, mdata.get());
 	if (!FRFXLL_SUCCESS(fxResData)) {
 		FRFXLLCloseHandle(&hFeatureSet);
 		throw NFIQ2::Exception(
@@ -197,9 +198,9 @@ NFIQ2::QualityFeatures::FingerJetFXFeature::computeFeatureData(
 
 	this->minutiaData_.clear();
 	this->minutiaData_.reserve(minCnt);
-	for (int i = 0; i < minCnt; i++) {
-		this->minutiaData_.emplace_back(
-		    static_cast<unsigned int>(mdata[i].x),
+	for (unsigned int i = 0; i < minCnt; i++) {
+		this->minutiaData_.emplace_back(static_cast<unsigned int>(
+						    mdata[i].x),
 		    static_cast<unsigned int>(mdata[i].y),
 		    static_cast<unsigned int>(mdata[i].a),
 		    static_cast<unsigned int>(mdata[i].q),
@@ -247,8 +248,8 @@ NFIQ2::QualityFeatures::FingerJetFXFeature::computeFeatureData(
 	rect200x200.height = 200;
 	vecRectDimensions.push_back(rect200x200);
 
-	FingerJetFXFeature::FJFXROIResults roiResults = computeROI(
-	    32, fingerprintImage, vecRectDimensions);
+	FingerJetFXFeature::FJFXROIResults roiResults = computeROI(32,
+	    fingerprintImage, vecRectDimensions);
 	double noOfMinInRect200x200 = 0;
 	for (unsigned int i = 0;
 	     i < roiResults.vecNoOfMinutiaeInRectangular.size(); i++) {
@@ -373,8 +374,10 @@ NFIQ2::QualityFeatures::FingerJetFXFeature::computeROI(int bs,
 
 		unsigned int noOfMinutiaeInRect = 0;
 		for (const auto &m : this->minutiaData_) {
-			if (m.x >= startX && m.x <= endX && m.y >= startY &&
-			    m.y <= endY) {
+			if (m.x >= static_cast<unsigned int>(startX) &&
+			    m.x <= static_cast<unsigned int>(endX) &&
+			    m.y >= static_cast<unsigned int>(startY) &&
+			    m.y <= static_cast<unsigned int>(endY)) {
 				// minutia is inside rectangular
 				noOfMinutiaeInRect++;
 			}
